Remove reintializing Ray code.
You can't keep shutting down and doing ray.init() in this module
because the cluster is going to start over again and you'll have a gap
in your time series metrics. This is to address the case where you want
to reference activity that happened in a previous exercise but can't
because you've cleared everything out.

Signed-off-by: Emmy Li <emmy@anyscale.com>