Ray Observability Module (#59)
* Initialize observability module Python file.

Signed-off-by: Emmy Li <emmy@anyscale.com>

* Initialize notebook for Ray Observability module.

Signed-off-by: Emmy Li <emmy@anyscale.com>

* Delete empty Python file.

Signed-off-by: Emmy Li <emmy@anyscale.com>

* Porting over just the code from draft document.

Signed-off-by: Emmy Li <emmy@anyscale.com>

* Remove reintializing Ray code.

You can't keep shutting down and doing ray.init() in this module
because the cluster is going to start over again and you'll have a gap
in your time series metrics. This is to address the case where you want
to reference activity that happened in a previous exercise but can't
because you've cleared everything out.

Signed-off-by: Emmy Li <emmy@anyscale.com>

* Minor formatting.

Signed-off-by: Emmy Li <emmy@anyscale.com>

* Updates to just the code.

Minor things including reinserting the reinitialization of clusters
which offers more controlled experience when viewing metrics in the
Dashboard. Some small fixes to errors. Swapping out an anti-pattern for
a hanging example in the debugging section.

Signed-off-by: Emmy Li <emmy@anyscale.com>

* Type hint for hang example.

Signed-off-by: Emmy Li <emmy@anyscale.com>

* Add "About this notebook" section

Signed-off-by: Emmy Li <emmy@anyscale.com>

* Add "Introduction to..." and "Ray obs..." sections

Added in "Introduction to observability in distributed systems" and
"Ray observability toolbox" sections. No images embedded yet.

Signed-off-by: Emmy Li <emmy@anyscale.com>

* Add "Debugging" subsection.

Signed-off-by: Emmy Li <emmy@anyscale.com>

* Add "Optimizing" and "Conclusion" subsections.

Signed-off-by: Emmy Li <emmy@anyscale.com>

* Add in diagrams.

Does not include screenshots yet.

Signed-off-by: Emmy Li <emmy@anyscale.com>

* Add in screenshots and adapt ray.get anti-pattern.

Two new screenshots and adapt the ray.get antipattern's big sleep time
to 25 seconds rather than 10. It needs to be longer than the longest
possible small sleep time.

Signed-off-by: Emmy Li <emmy@anyscale.com>

* Update README and rename module to match.

Signed-off-by: Emmy Li <emmy@anyscale.com>

---------

Signed-off-by: Emmy Li <emmy@anyscale.com>